Rename ExpandableInput to ExpanableOverlay
